mac80211: tell drivers the user TX power restriction
Allow drivers to obtain the user's TX power restrictions in
order to be able to  apply it properly on (for example)
hardware scanning which operations on a set of channels,
in which case mac80211 cannot properly apply the per-channel
regulatory limits, and the currently exposed txpower value,
which is the minimum  of the user setting, channel regulatory
and possible 11h AP reduction isn't  appropriate to use
since scanning is done on many channels.

Note that drivers need to be careful when using this value
and must ensure that  regulatory is maintained, i.e. the
new value cannot be used as-is.

Change-Id: Icf442cfa835cdbee71fe619b15f63b5d17a34c39
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
reapplied after merge
Reviewed-on: https://gerrit.rds.intel.com/r/51855
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: 09f3c196d54865100835d99204c08763a0108824