Revert "iwlwifi: sdio: support secured boot flow for family 8000 B step" [NOUPSTREAM]
This reverts commit 25add08ad9a402357a9e8cb30e2590b6a4b440eb.

The patch this reverts causes the driver to load the FW
before the DBGC is configured in SDIO mode, thus causing
the FW to NOT configure the DBGC, preventing an option for
debugging.

This revertion is until a fix is provided for this issue.

Change-Id: I78ab36103354a2a8cbf1462c21fb6f81267efd32
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/52732
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 8d81c6c00560f9bcc70af300428717d96a546a78