Revert "iwlwifi: pcie: support secured boot flow for family 8000 B step" [NOUPSTREAM]
This reverts commit ac3848414981d4c22469ab8a4b8d6ba672d31695.

The patch this reverts causes the driver to load the FW
before the DBGC is configured in PCIe mode, thus causing
the FW to NOT configure the DBGC, preventing an option for
debugging.

This revertion is until a fix is provided for this issue.

Change-Id: I1d221213b888fefd6b82d344b72e4638245ef1cf
Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/52733
Tested-by: IWL Jenkins
Reviewed-by: Dor Shaish <dor.shaish@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 65bcbbf8e2c37226f372be012e125aefaad1697a