[NOUPSTREAM]  mac80211: Tx latency: Add more timestamps to the latency statistics
Currently the Tx latency is calculated as the delta between 2 timestamps,
the time the skb entered the kernel, and the time it is about to be
deleted.

In order to improve the analysis of the Tx latency, it is crucial to know
what part of the Tx flow is causing the high latency.

Add option to support additional timestamps, the time the skb is written
to the bus, and the time the skb is acked by the AP.

Add option to change the 2 timestamps that are used to calculate the
measurements, in order to be able to understand where in the Tx flow is
the problematic part.

Change-Id: I66454f2d48b616b2014b93a0023767997183237b
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/53062
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: ef43eaa71abafc247f4dd155ed01f4f79857fc91