iwlwifi: mvm: init U-APSD non-agg detection rate before association
If we just initialize the average rate once, then it can 'leak' from
one association to the next, causing issues. Instead of initialising
it once, do it before the association. This will also reinit it when
firmware restart happens, but that might actually be a good thing as
throughput will be low in that case.

Change-Id: Ia5677bb6aa5c6b7ea02daef7409ba516bc2d6cdb
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/53127
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: d5628c919f4259427beea539ca2db0c0ff664a34