mac80211: Adjust chan_ctx when assigning reserved vif[BUGFIX]
It is possible to reserve a narrower channel context during csa (as long as
the chandefs are compatible). However when the reservation is used, the
channel context should be adjusted. Not doing so would result in attempts
to use higher rates than the chan_ctx supports.
---
MWG100229683

Change-Id: Ia403a6f48a3aa7fcdda9dbe26b9f6e6326735e86
Signed-off-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/53431
Reviewed-by: Luciano Coelho <luciano.coelho@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 66d9b55405152881786477c9b433eb69fe27a04d