iwlwifi: mvm: allow to configure the timeout for the Tx queues
Sometimes we will want to configure the timeouts for the
Tx queues based on the vif type. Allow to do that using the
trigger mechanism.

type=feature

Change-Id: I1bf09ddf97af7aad0a08e8ffd9d4d986951223cb
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
Golan - for this trigger, we should not allow to set the vif_type
since it is already in the trigger itself. It should be clear from
the GUI that the timeouts are in milliseconds and that 0 means to
disable to the timer completely.
The default should be 10000ms for all the queues, besides the command
queue which has a default timeout of 2500ms
Reviewed-on: https://gerrit.rds.intel.com/r/54306
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: d70262444ac490851e93eb31a46fab7d9cc7348f