iwlwifi: add new TLV capability flag for gscan support
Gscan is a scan feature which is supported on certain devices only,
hence the need for a TLV flag for it. For devices that support gscan
store the gscan capabilities advertised by the FW so the driver can
report it to upper layers.

type=feature

Change-Id: Iab6e49b7b15b668b60aa9baa57546b4c094270d7
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/54358
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 06ab2a28298e579bf92037e925c3265aeae54fb1