iwlwifi: idi: enable DBGM
Enable DBGM over IDI (as rx frames, using out of band channel)

There are 2 debug buffers. Each one has associated event,
that when invoked will configure the kick2 (through LL)
to pass the relevant debug buffer to the rx fifo.

In order to use the debug buffers, we have to decrease
the TBs count. In order to minimize the changes to the
IDI layer, redefine some macros (which used to have
a const value) to point to a "layout struct", that
can be changed in runtime through the ini file.

type=feature

Change-Id: I859e9b28e2d0b7856c1b072a2f4993a0cf9f8e3b
Signed-off-by: Eliad Peller <eliadx.peller@intel.com>
Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Signed-off-by: Idan Kahlon <idanx.kahlon@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/55341
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Guy Cohen <guy.cohen@intel.com>
x-iwlwifi-stack-dev: d51eda2fb19c37d56444231bcdd8ac975b72c141