iwlwifi: pcie: support frag SKBs
Allow frag SKBs in PCIe and advertise the maximum number of frags
to the opmode. As a fallback. linearize the SKB if it exceeds the
maximum number of fragments. This allows using the hardware better
(filling more TBs) and should improve performance when used by the
opmode.

Also adjust tracing to be able to deal with this.

type=feature

Change-Id: I4c5e4acde4ebd88f98e12d78e5281946371b575c
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/59249
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: e43d64d1b65b9d4497765b5bc04130c79688cd6f