iwlwifi: mvm: change the platform nvm file selection flow
Target of the new flow is to allow lnP (sdio and pcie) to set default
nvm file. in case that userspace specify nvm file that doesn't match
to the HW the driver will replace it with the default.
I added this patch since that there are some pcie LnP platforms.

The new flow:
1. in case that "mod_param.nvm_file" isn't NULL driver takes this file
as default and try to parse it.
2. in case that the "mod_param.nvm_file" doesn't match to the HW step
driver takes another file from "iwl_cfg.default_nvm_file" and try to work
with the new file.
3. in case that "mod_param.nvm_file" is NULL driver doesn't load any
external file.

type=feature

Change-Id: Iedc34bd866452aa1cc2f95266abab71c26bfeed9
Signed-off-by: Eran Harary <eran.harary@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/55494
Tested-by: IWL Jenkins
Reviewed-by: Dor Shaish <dor.shaish@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 05aa0c87702999015c57bf8545496569a1bb52f6