[BUGFIX] mac80211: use the BSS chandef for HT/VHT operation IEs
The chandef of the current channel context might be wider (though
compatible). The TDLS link cares only about the channel of the BSS.

In addition make sure to specify the VHT operation IE when VHT is supported
on a non-2.4GHz band, as required by IEEE802.11ac-2013. This is not the
same as HT-operation, to be specified only if the BSS doesn't support HT.

type=bugfix
bug=not-tracked
fixes=unknown

Change-Id: I0d1b439c0bddfeb3edf41c0b4004cdb91198f1e0
Signed-off-by: Arik Nemtsov <arikx.nemtsov@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/56370
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: fc53f8a552642287bd2664edf804a70c20706d55