iwlwifi: mvm: Frequency Manager: Handle invalid mitigation channel list
Validate the list of channels that the FM requests to apply
mitigations on.

Make sure that no channel in the list appears twice and that
each channel in the list is currently in use.

If the list is not valid - return an error status.

type=cleanup

Change-Id: I7794962d014e012034b0362d65c299a86f105d6d
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/55186
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: e92ecdbdbdeb0de79ec6e8b03c36ec003d0f77c6