[BUGFIX][NOUPSTREAM] mac80211: Tx latency: Avoid kernel panic
The points variable, which enables the tx latency statistics tool to
measure the latency time for tx packet, is defined in the tx
latency struct.
So when one uses the tx consecutive loss mechanism but not the tx latency
statistics tool, the points are used but not allocated and a kernel panic
may occur.

Remove the points variable from the tx latency struct and allocate
them statically in order to solve this.

In Addition, this could be useful for other mechanisms which would like
to perform tx latency measurements.

type=bugfix
bug=not-tracked
fixes=I66454f2d48b616b2014b93a0023767997183237b

Change-Id: I23fb7a33f0248bf7628c6e9a757dc3fd7a5e978d
Signed-off-by: Golan Ben Ami <golan.ben.ami@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/57310
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 0498d87dbea2225f4d2ebe87419ee3646768cab6