mac80211: Add support for declaring MU-MIMO capability
Add support for declaring MU-MIMO beamformee capability for
relevant hardware.
When sending association request, the capability is included if both
hardware and the AP support it, and no other virtual interface
is using it.
This is in order to avoid multiple interfaces using MU-MIMO in parallel
which might lead to contradictions in the group-id mechanism.

type=feature

Change-Id: Ic7ec12b62bf46a2966caebe35b9bcbd42e1b4948
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/57722
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
x-iwlwifi-stack-dev: b00ca68ca65767b849ba103a36cd6e16e2295709