iwlwifi: mvm: Frequency Manager: Add bit mask to notify/mitigate API
The FM can request to mitigate different things in one request,
but not always will the FM want to apply all mitigations in the
same request. There is no straight forward way to determine what
are the exact mitigations that need to be applied.
The more mitigations that will be added in the future the harder it
will get to determine which mitigation FM wants to apply.

The driver currently notifies the FM regarding changes in channel
activity, but in the future more notifications will be added and it
will be possible to notify the FM regarding a few notifications
in the same notification. The more notifications that will be
added the harder it will get for the FM determine which notification
the driver wants to notify about.

Add a bit mask to the API between the FM-driver to indicate the
exact mitigation/notifications that are taking place.

In addition add a small cleanup to the ut module (fm-test.c) that
that is required due to the new API change.

[Note the API is not formal yet - so the names of the parameters might
change - but i pushed it to be reviewd in the mean-time]

type=cleanup

Change-Id: Ia60edfbdd8171fa117af89a1c4ce8c7959a8893e
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/57829
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 48e5dd4fbcee4c89616e8585a3ff6783e61274bf